# src/docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "TeamTime API",
    "version": "1.0.0",
    "description": "API REST para gestión de tiempo, proyectos y recursos humanos con sistema de roles jerárquico",
    "contact": {
      "name": "TeamTime Development Team",
      "email": "desarrollo@teamtime.com",
      "url": "https://github.com/your-org/teamtime-backend"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Servidor de desarrollo"
    },
    {
      "url": "https://api.teamtime.com/api",
      "description": "Servidor de producción"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token JWT obtenido del endpoint /auth/login"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "success",
          "message"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "Descripción del error"
          },
          "error": {
            "type": "string",
            "example": "VALIDATION_ERROR"
          },
          "details": {
            "type": "object",
            "properties": {
              "field": {
                "type": "string",
                "example": "email"
              },
              "code": {
                "type": "string",
                "example": "INVALID_FORMAT"
              }
            }
          }
        }
      },
...
      "Project": {
        "type": "object",
        "required": [
          "id",
          "name",
          "areaId"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "project-123"
          },
          "name": {
            "type": "string",
            "example": "Sistema de Gestión"
          },
          "description": {
            "type": "string",
            "example": "Desarrollo del sistema de gestión empresarial"
          },
          "areaId": {
            "type": "string",
            "format": "uuid",
            "example": "area-456"
          },
          "status": {
            "type": "string",
            "enum": [
              "PLANIFICADO",
              "EN_PROGRESO",
              "COMPLETADO",
              "CANCELADO"
            ],
            "example": "EN_PROGRESO"
          },
          "startDate": {
            "type": "string",
            "format": "date",
            "example": "2025-07-01"
          },
          "endDate": {
            "type": "string",
            "format": "date",
            "example": "2025-12-31"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2025-07-04T10:30:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2025-07-04T10:30:00.000Z"
          }
        }
      },
...
    },
    "responses": {
      "UnauthorizedError": {
        "description": "Token de autenticación inválido o faltante",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "success": false,
              "message": "Token de autenticación requerido",
              "error": "UNAUTHORIZED"
            }
          }
        }
      },
      "ForbiddenError": {
        "description": "Permisos insuficientes para realizar la operación",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "success": false,
              "message": "No tienes permisos para realizar esta acción",
              "error": "FORBIDDEN"
            }
          }
        }
      },
      "NotFoundError": {
        "description": "Recurso no encontrado",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "success": false,
              "message": "Recurso no encontrado",
              "error": "NOT_FOUND"
            }
          }
        }
      },
      "ValidationError": {
        "description": "Error de validación en los datos enviados",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "success": false,
              "message": "Error de validación",
              "error": "VALIDATION_ERROR",
              "details": {
                "field": "email",
                "code": "INVALID_FORMAT"
              }
            }
          }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/areas/{id}/projects": {
      "get": {
        "summary": "Obtener proyectos de un área",
        "description": "Retorna la lista de proyectos pertenecientes a un área específica.",
        "tags": [
          "Áreas"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "ID del área"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            },
            "description": "Número de página"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 10
            },
            "description": "Cantidad de elementos por página"
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "PLANIFICADO",
                "EN_PROGRESO",
                "COMPLETADO",
                "CANCELADO"
              ]
            },
            "description": "Filtrar por estado del proyecto"
          }
        ],
        "responses": {
          "200": {
            "description": "Proyectos del área obtenidos exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Proyectos del área obtenidos exitosamente"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "projects": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Project"
                          }
                        },
                        "pagination": {
                          "type": "object",
                          "properties": {
                            "page": {
                              "type": "integer",
                              "example": 1
                            },
                            "limit": {
                              "type": "integer",
                              "example": 10
                            },
                            "total": {
                              "type": "integer",
                              "example": 15
                            },
                            "pages": {
                              "type": "integer",
                              "example": 2
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          }
        }
      }
    },
...
    "/projects": {
      "post": {
        "summary": "Crear nuevo proyecto",
        "description": "Permite a administradores y coordinadores crear un nuevo proyecto en el sistema.",
        "tags": [
          "Proyectos"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "areaId"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Sistema de Gestión CRM"
                  },
                  "description": {
                    "type": "string",
                    "example": "Desarrollo de sistema CRM para gestión de clientes"
                  },
                  "areaId": {
                    "type": "string",
                    "format": "uuid",
                    "example": "area-123"
                  },
                  "startDate": {
                    "type": "string",
                    "format": "date",
                    "example": "2025-07-01"
                  },
                  "endDate": {
                    "type": "string",
                    "format": "date",
                    "example": "2025-12-31"
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "PLANIFICADO",
                      "EN_PROGRESO",
                      "COMPLETADO",
                      "CANCELADO"
                    ],
                    "default": "PLANIFICADO",
                    "example": "PLANIFICADO"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Proyecto creado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Proyecto creado exitosamente"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Project"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          }
        }
      },
      "get": {
        "summary": "Obtener lista de proyectos",
        "description": "Retorna una lista paginada de proyectos con filtros opcionales. Los usuarios ven proyectos según sus permisos.",
        "tags": [
          "Proyectos"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            },
            "description": "Número de página"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 10
            },
            "description": "Cantidad de elementos por página"
          },
          {
            "in": "query",
            "name": "areaId",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Filtrar por área"
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "PLANIFICADO",
                "EN_PROGRESO",
                "COMPLETADO",
                "CANCELADO"
              ]
            },
            "description": "Filtrar por estado del proyecto"
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            },
            "description": "Buscar por nombre o descripción"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de proyectos obtenida exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Proyectos obtenidos exitosamente"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "projects": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Project"
                          }
                        },
                        "pagination": {
                          "type": "object",
                          "properties": {
                            "page": {
                              "type": "integer",
                              "example": 1
                            },
                            "limit": {
                              "type": "integer",
                              "example": 10
                            },
                            "total": {
                              "type": "integer",
                              "example": 45
                            },
                            "pages": {
                              "type": "integer",
                              "example": 5
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        }
      }
    },
    "/projects/{id}": {
      "get": {
        "summary": "Obtener proyecto por ID",
        "description": "Retorna la información detallada de un proyecto específico.",
        "tags": [
          "Proyectos"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "ID del proyecto"
          }
        ],
        "responses": {
          "200": {
            "description": "Proyecto obtenido exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Proyecto obtenido exitosamente"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Project"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          }
        }
      }
    },
...
  }
}